backend/requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0
alembic==1.13.1
psycopg2-binary==2.9.9

